CP-1467 #Device&Tests First Version

+++ app/Http/Controllers/DevicesController.php
@@ -6,11 +6,16 @@ use Cartalyst\DataGrid\Laravel\Facades\DataGrid;
 use Dingo\Api\Routing\Helpers;
 use Illuminate\Session\SessionManager as Session;
 use WA\DataStore\Device\DeviceTransformer;
+use WA\DataStore\Device\Device;
 use WA\Helpers\Traits\SetLimits;
 use WA\Http\Controllers\Api\Traits\BasicCrud;
 use Illuminate\Http\Request;
 
 use WA\Repositories\Device\DeviceInterface;
+
+use Validator;
+use DB;
+use Log;
 
 /**
  * Devices resource.
@@ -62,8 +67,12 @@ class DevicesController extends ApiController
      */
     public function show($id)
     {
-        $device = $this->device->byId($id);
-
+        $device = Device::find($id);
+        if($device == null){
+            $error['errors']['get'] = 'the Device selected doesn\'t exists';   
+            return response()->json($error)->setStatusCode(409);
+        }
+        
         return $this->response()->item($device, new DeviceTransformer(),['key' => 'devices']);
     }
 
@@ -102,10 +111,95 @@ class DevicesController extends ApiController
      */
     public function store($id, Request $request)   
     {
-        $data = $request->all();       
-        $data['id'] = $id;
-        $device = $this->device->update($data);
-        return $this->response()->item($device, new DeviceTransformer(), ['key' => 'devices']);
+        $data = $request->all()['data'];
+        $success = true;
+
+        $dataType = $data['type'];
+        $dataAttributes = $data['attributes'];
+        $dataRelationships = $data['relationships'];
+        $dataAssets = $this->parseJsonToArray($dataRelationships['assets']['data']);
+        $dataModifications = $this->parseJsonToArray($dataRelationships['modifications']['data']);
+        $dataCarriers = $this->parseJsonToArray($dataRelationships['carriers']['data']);
+        $dataCompanies = $this->parseJsonToArray($dataRelationships['companies']['data']);
+        $dataPrices = $dataRelationships['prices']['data'];
+
+        DB::beginTransaction();
+
+        $validator = Validator::make(array('image' => $dataAttributes['image']), ['image' => 'url']);
+        if($validator->fails()){
+            $error['errors']['image'] = 'The image file has not a valid format';
+        }
+        
+        try{
+            $device = Device::find($id);
+
+            $device->image = $dataAttributes['image'];
+            $device->name = $dataAttributes['name'];
+            $device->properties = $dataAttributes['properties'];
+            $device->deviceTypeId = $dataAttributes['deviceTypeId'];
+
+            $device->save();
+            
+        } catch (\Exception $e) {
+            DB::rollBack();
+            $success = false;
+            $error['errors']['devices'] = 'The Device can not be updated';
+            return response()->json($error)->setStatusCode(409);
+        }
+
+        try {
+            $device->assets()->sync($dataAssets);    
+        } catch (\Exception $e){
+            DB::rollBack();
+            $success = false;
+            $error['errors']['assets'] = 'the Device Assets can not be updated';
+        }
+
+        try {
+            $device->modifications()->sync($dataModifications);
+        } catch (\Exception $e){
+            DB::rollBack();
+            $success = false;
+            $error['errors']['modifications'] = 'the Device Modifications can not be updated';
+        }
+
+        try {
+            $device->carriers()->sync($dataCarriers);
+        } catch (\Exception $e){
+            DB::rollBack();
+            $success = false;
+            $error['errors']['carriers'] = 'the Device Carriers can not be updated';
+        }
+
+        try {
+            $device->companies()->sync($dataCompanies);
+        } catch (\Exception $e){
+            DB::rollBack();
+            $success = false;
+            $error['errors']['companies'] = 'the Device Companies can not be updated';
+        }
+        
+        try{
+            $priceInterface = app()->make('WA\Repositories\Device\DevicePriceInterface');
+
+            foreach ($dataPrices as $price) {
+                $price['deviceId'] = $device->id;
+                $priceInterface->update($price);
+            }    
+
+        } catch (\Exception $e) {
+            DB::rollBack();
+            $success = false;
+            $error['errors']['prices'] = 'the Device Prices can not be updated';
+        }
+
+        if(!$success){
+            DB::rollBack();
+            return response()->json($error)->setStatusCode(409);
+        } else {
+            DB::commit();
+            return $this->response()->item($device, new DeviceTransformer(), ['key' => 'devices']);
+        }       
     }
 
     /**
@@ -114,27 +208,95 @@ class DevicesController extends ApiController
      * @return \Dingo\Api\Http\Response
      */
     public function create(Request $request)
-    {
-        $data = $request->all();
-        $dataDevice = [
-            'image' => $data['image'],
-            'name' => $data['name'],
-            'properties' => $data['properties']
-            ];
-        var_dump($dataDevice);
-        $device = $this->device->create($dataDevice);
-        var_dump($device->id);
-        die;
+    {        
+        $data = $request->all()['data'];
+        $success = true;
 
+        $dataType = $data['type'];
+        $dataAttributes = $data['attributes'];
+        $dataRelationships = $data['relationships'];
+        $dataAssets = $this->parseJsonToArray($dataRelationships['assets']['data']);
+        $dataModifications = $this->parseJsonToArray($dataRelationships['modifications']['data']);
+        $dataCarriers = $this->parseJsonToArray($dataRelationships['carriers']['data']);
+        $dataCompanies = $this->parseJsonToArray($dataRelationships['companies']['data']);
+        $dataPrices = $dataRelationships['prices']['data'];
+        
+        DB::beginTransaction();
 
-        $modification = app()->make('WA\Repositories\Device\DeviceModificationInterface');
-        //$dataModification = ['deviceId' => $device->id, 'modificationId']
-        //$modification->create();
+        $validator = Validator::make(array('image' => $dataAttributes['image']), ['image' => 'url']);
+        if($validator->fails()){
+            $error['errors']['image'] = 'The image file has not a valid format';
+        }
 
-        $carrier = app()->make('WA\Repositories\Device\DeviceCarrierInterface');
-        $company = app()->make('WA\Repositories\Device\DeviceCompanyInterface');
-        $price = app()->make('WA\Repositories\Device\DevicePriceInterface');
-        return $this->response()->item($device, new DeviceTransformer(), ['key' => 'devices']);
+        try{
+            $device = $this->device->create($dataAttributes);
+
+        } catch (\Exception $e) {
+            Log::info("Exception dataAttributes: ".var_dump($e));
+            DB::rollBack();
+            $success = false;
+            $error['errors']['devices'] = 'The Device can not be created';
+            return response()->json($error)->setStatusCode(409);
+        }
+
+        try {
+            $device->assets()->sync($dataAssets);    
+        } catch (\Exception $e){
+            Log::info("Exception dataAssets: ".var_dump($e));
+            DB::rollBack();
+            $success = false;
+            $error['errors']['assets'] = 'the Device Assets can not be created';
+        }
+
+        try {
+            $device->modifications()->sync($dataModifications);
+        } catch (\Exception $e){
+            Log::info("Exception dataModifications: ".var_dump($e));
+            DB::rollBack();
+            $success = false;
+            $error['errors']['modifications'] = 'the Device Modifications can not be created';
+        }
+
+        try {
+            $device->carriers()->sync($dataCarriers);
+        } catch (\Exception $e){
+            Log::info("Exception dataCarriers: ".var_dump($e));
+            DB::rollBack();
+            $success = false;
+            $error['errors']['carriers'] = 'the Device Carriers can not be created';
+        }
+
+        try {
+            $device->companies()->sync($dataCompanies);
+        } catch (\Exception $e){
+            Log::info("Exception dataCompanies: ".var_dump($e));
+            DB::rollBack();
+            $success = false;
+            $error['errors']['companies'] = 'the Device Companies can not be created';
+        }
+        
+        try{
+            $priceInterface = app()->make('WA\Repositories\Device\DevicePriceInterface');
+
+            foreach ($dataPrices as $price) {
+                $price['deviceId'] = $device->id;
+                $priceInterface->create($price);
+            }    
+
+        } catch (\Exception $e) {
+            Log::info("Exception dataPrices: ".var_dump($e));
+            DB::rollBack();
+            $success = false;
+            $error['errors']['prices'] = 'the Device Prices can not be created';
+        }
+
+        if(!$success){
+            DB::rollBack();
+            return response()->json($error)->setStatusCode(409);
+        } else {
+            DB::commit();
+            return $this->response()->item($device, new DeviceTransformer(), ['key' => 'devices']);
+        }
     }
 
     /**
@@ -144,7 +306,31 @@ class DevicesController extends ApiController
      */
     public function delete($id)
     {
-        $this->device->deleteById($id);
+        $device = Device::find($id);
+        if($device <> null){
+            $this->device->deleteById($id);
+        } else {
+            $error['errors']['delete'] = 'the Device selected doesn\'t exists';   
+            return response()->json($error)->setStatusCode(409);
+        }
+        
         $this->index();
+        $device = Device::find($id);        
+        if($device == null){
+            return array("success" => true);
+        } else {
+            $error['errors']['delete'] = 'the Device has not been deleted';   
+            return response()->json($error)->setStatusCode(409);
+        }
+    }
+
+    private function parseJsonToArray($data){
+        $array = array();
+        
+        foreach ($data as $info) {
+            array_push($array, $info['id']);
+        }
+        
+        return $array;
     }
 }
